requirements.txt
fastapi==0.115.12
jinja2==3.1.6
pip-chill==1.0.3
python-dotenv==1.0.1
uvicorn==0.34.0
pydantic-settings==2.8.1
langchain==0.3.22
langgraph==0.3.29
fastapi[standard]
langchain-openai
langchain_groq